requirements.txt
ultralytics==8.1.2
streamlit==1.31.1
torch>=2.0.0
torchvision>=0.15.0
opencv-python-headless==4.9.0.80
numpy>=1.24.0
pillow>=10.0.0
pandas>=2.0.0
pyyaml>=6.0.0
requests>=2.31.0
matplotlib>=3.8.0
seaborn>=0.13.0
tqdm>=4.66.0 